New script: YouTube: TV-sized banner toggle
This script makes it easy to see the TV-sized banners on YouTube channels.
Simply click on the banner on the YouTube channel to toggle it between regular
size and TV size.

If AdBlock Plus is used, an exception rule will need to be added as detailed in
the source code, otherwise it may show blank images for some channels.