New script: Remove ?utm_* attributes
These attributes are, in my opinion, the scourge of the Web. They're
used by web analytics software for marketing purposes, and are used to
track inbound links above and beyond simply using the Referer HTTP
header.

This particular script also includes a last line of defense: When you
click a link, it will catch the click as late as possible in the chain
and re-check it for these attributes. It will then clean any if found
before passing it off to the browser.