New script: Reddit: Stop invasive UTM tracking
This script prevents Reddit's UTM tracking from happening when you click
a self-hosted link.

While I do have a last line of defense script that works for Reddit (and
other sites) in most cases (see the "Remove UTM attributes" script in
this repository), it's not exactly clean, and it fails to work for
Ctrl-clicking to open links in a new tab. This script is a much cleaner
script targetting Reddit specifically.