New script: Allow speaking of IPA on Wikitionary
...and also Wikipedia, though I couldn't fit it into the commit title.

This script auto-links the IPA on Wikitionary and Wikipedia, such that
when you click the links, your browser will speak the IPA! It doesn't
work for everything, though - in particular, it only supports 'broad'
IPA and not 'narrow' IPA.

This commit also includes the resource needed to use this script,
resources/mespeak.js.

This is the first script to the repository that is released under the
General Public License, version 3.0. This is necessary because the
original code was released under the GPL. The LICENSE file has been
modified accordingly.