Query string parser: Allow empty string argument
I wanted to be able to use this parser on a link's search component
today, but it turns out that the parser defaults to the current page's
query string if no argument is passed in. Unfortunately, the test breaks
if you *do* pass in an argument but it's an empty string.

Thus, we now use a stricter method of testing for definedness on the
argument that's passed in.

This commit is dedicated to the memory of Jan Wolter, the original
author of this decoder, who died in 2015 from complications of
influenza. (https://unixpapa.com/)